feat: add stream token proxy
